git-svn-id: https://10.0.16.5/svn/tshow/server/meShow_new/branches/meShow_beta@104850 2fe1c438-e5a1-4f48-b1bb-90b7b6eba2af

+++ src/main/java/com/melot/kktv/domain/ConfigGiftInfo.java
@@ -21,7 +21,7 @@ public class ConfigGiftInfo {
     /**
      * 配置信息总的版本号
      */
-    Long giftVersion;
+    Integer giftVersion;
     
     /**
      * 资源信息配置信息
@@ -38,11 +38,11 @@ public class ConfigGiftInfo {
      */
     List<GiftIconUrl> iconUrl;
 
-    public Long getGiftVersion() {
+    public Integer getGiftVersion() {
         return giftVersion;
     }
 
-    public void setGiftVersion(Long giftVersion) {
+    public void setGiftVersion(Integer giftVersion) {
         this.giftVersion = giftVersion;
     }
